requirements.txt
Pillow
sentence_transformers
elasticsearch<9.0.0
matplotlib
